+++ src/api/main.py
@@ -3,13 +3,12 @@
 
 from fastapi import FastAPI
 
-from api.routers import task, done
+from api.routers import users
 
 app = FastAPI()
 
 
-app.include_router( task.router )
-app.include_router( done.router )
+app.include_router( users.router )
 
 
 # End of Script
